Upgrade to latest pytype, make integration tests run only when python changes, and only pytype/pylint changed files.

+++ run-travis-test.sh
@@ -1,6 +1,18 @@
 #!/bin/bash
 
-if [ "${MATRIX_SHARD}" = "sanity" ] ; then
+
+echo TRAVIS_COMMIT_RANGE: $TRAVIS_COMMIT_RANGE
+echo TRAVIS_COMMIT: $TRAVIS_COMMIT
+FILES_CHANGED=`git diff --name-only $TRAVIS_COMMIT_RANGE`
+PY_FILES_CHANGED=`git diff --name-only $TRAVIS_COMMIT_RANGE|grep -E ".py$"`
+
+if [[ "$FILES_CHANGED" != "" ]] ; then
+  echo files changed: $FILES_CHANGED
+else
+  echo no files changed.
+fi
+
+if [ "${MATRIX_SHARD}" == "sanity" ] ; then
   FAUCET_TESTS="-u FaucetSanityTest"
   ./tests/run_unit_tests.sh || exit 1
   codecov || true
@@ -30,6 +42,11 @@ if [[ "$PY3V" != "Python 3.6"* ]] ; then
   exit 0
 fi
 
+if [[ "$PY_FILES_CHANGED" == "" ]] ; then
+  echo no python source changed, not running docker tests.
+  exit 0
+fi
+
 docker pull faucet/test-base
 docker build -t ${FAUCET_TEST_IMG} -f Dockerfile.tests . || exit 1
 docker rmi faucet/test-base
@@ -39,5 +56,6 @@ echo Shard $MATRIX_SHARD: $FAUCETTESTS
 sudo docker run --privileged --sysctl net.ipv6.conf.all.disable_ipv6=0 \
   -v $HOME/.cache/pip:/var/tmp/pip-cache \
   -e FAUCET_TESTS="${FAUCET_TESTS}" \
+  -e PY_FILES_CHANGED="${PY_FILES_CHANGED}" \
   -t ${FAUCET_TEST_IMG} || exit 1
 exit 0
